authorization implemented

+++ src/api/notes/handler.js
@@ -13,7 +13,8 @@ class NotesHandler {
   async postNoteHandler (request, h) {
     this._validator.validateNotePayload(request.payload)
     const { title = 'untitled', body, tags } = request.payload
-    const noteId = await this._service.addNote({ title, body, tags })
+    const { id: credentialId } = request.auth.credentials
+    const noteId = await this._service.addNote({ title, body, tags, owner: credentialId })
     const response = h.response({
       status: 'success',
       message: 'Catatan berhasil ditambahkan',
@@ -25,8 +26,9 @@ class NotesHandler {
     return response
   }
 
-  async getNotesHandler () {
-    const notes = await this._service.getNotes()
+  async getNotesHandler (request) {
+    const { id: credentialId } = request.auth.credentials
+    const notes = await this._service.getNotes(credentialId)
     return {
       status: 'success',
       data: {
@@ -37,6 +39,8 @@ class NotesHandler {
 
   async getNoteByIdHandler (request) {
     const { id } = request.params
+    const { id: credentialId } = request.auth.credentials
+    await this._service.verifyNoteOwner(id, credentialId)
     const note = await this._service.getNoteById(id)
     return {
       status: 'success',
@@ -50,6 +54,8 @@ class NotesHandler {
     this._validator.validateNotePayload(request.payload)
     const { title, body, tags } = request.payload
     const { id } = request.params
+    const { id: credentialId } = request.auth.credentials
+    await this._service.verifyNoteOwner(id, credentialId)
 
     // this._service.editNoteById(id, request.payload);
     await this._service.editNoteById(id, { title, body, tags })
@@ -62,7 +68,8 @@ class NotesHandler {
 
   async deleteNoteByIdHandler (request) {
     const { id } = request.params
-    // this._service.deleteNoteById(id);
+    const { id: credentialId } = request.auth.credentials
+    await this._service.verifyNoteOwner(id, credentialId)
     await this._service.deleteNoteById(id)
     return {
       status: 'success',
